Add javadocs to no.digipost.util.bisect
